src/requirements.txt
azure-core==1.21.0
azure-identity==1.7.1
azure-storage-queue==12.1.6
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cryptography==36.0.0
Flask==1.1.4
google-api-core==2.2.2
google-auth==2.3.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.3
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.0
opencensus-ext-flask==0.7.6
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.7.6
portalocker==1.7.1
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.13.3
